+++ HISTORY.rst
@@ -4,6 +4,15 @@ History
 0.2 (In development)
 --------------------
 
+Changed:
+
+* Added a ``markus.backends.logging.LoggingRollupMetrics`` backend that
+  rolls up metrics and does some light math on them. Possibly helpful
+  for light profiling for development.
+
+* Lots of documentation fixes. Thank you, Peter!
+
+
 0.1 (April 10th, 2017)
 ----------------------